+++ CHANGELOG.md
@@ -8,6 +8,7 @@ they can and will change without that change being reflected in Styler's semanti
 ### Fixes
 
 * fix `with` arrow replacement + redundant body removal creating invalid statements (#184, h/t @JesseHerrick)
+* allow Kernel unary `!` and `not` as valid pipe starts (#183, h/t @nherzing)
 
 ## 1.0.0-rc.2